+++ src/presentation/routes/hello_apps_route.py
@@ -4,17 +4,17 @@ import pythonbible as bible
 
 from src.presentation.handler.auth import verify_api_key
 from src.presentation.handler.hello_word_handler import get_random_verse
-from ..handler.responses import MyResponse
+from ..handler.responses import MyResponseModel, MyResponse
 
 hello_apps_v1 = APIRouter(prefix='/v1')
 
-@hello_apps_v1.get("/hello_apps", tags=["Apps"], 
+@hello_apps_v1.get("/hello_apps", 
                     summary="Hello Apps Endpoint",
                     description="This endpoint returns a random verse from the Bible.",
-                    response_model=MyResponse)
+                    response_model=MyResponseModel)
 def hello_apps():
     return MyResponse(
-        status=status.HTTP_200_OK,
+        status_code=status.HTTP_200_OK,
         message="Hello, World, this is an opened test endpoint.",
         data=get_random_verse(bible.Version.KING_JAMES),
     )