+++ src/presentation/schema/apps_schema.py
@@ -2,6 +2,7 @@
 from abc import abstractmethod
 from dataclasses import dataclass
 from src.presentation.routes.hello_apps_route import hello_apps_v1
+from src.presentation.routes.exchange_app_route import exchange_app_route
 
 
 @abstractmethod
@@ -12,6 +13,7 @@ class AppsBaseSchema:
     """
 
     path_segment: str = "/apps"
+    tag: str = "Apps"
 
     @property
     @abstractmethod
@@ -35,11 +37,13 @@ class AppsBaseSchema:
 class AppsRoutes:
 
     hello_apps_v1 = hello_apps_v1
+    exchange_app_route = exchange_app_route
 
 
 class AvailableApps:
 
     api = "/api"
+
 
 
 class AppsSchema(AppsBaseSchema):